Ensuring that all of the code is being refactored correctly - with all of the operations
